requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
gunicorn==23.0.0
Jinja2==3.1.6
pydantic==2.11.10
pyrate-limiter==3.9.0
python-dotenv==1.1.1
